[Silver III] Title: 주유소, Time: 420 ms, Memory: 35740 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -6,8 +6,8 @@ public class Main {
         BufferedReader br =new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
         StringTokenizer st= new StringTokenizer(br.readLine());
-        int[] distance = new int[N-1];
-        int[] price = new int[N];
+        long[] distance = new long[N-1];
+        long[] price = new long[N];
         for(int i=0;i<N-1;i++){
             distance[i]= Integer.parseInt(st.nextToken());
         }
@@ -16,17 +16,17 @@ public class Main {
             price[i] = Integer.parseInt(st.nextToken());
         }
 
-        int ans =0;
-        int currentPick = price[0];
-        ans+= (currentPick*distance[0]);
+        long ans =0;
+        long currentPick = price[0];
+        ans+= ( currentPick *distance[0]);
 
 
         for(int i=1;i<N-1;i++){
             if((currentPick*distance[i])>(price[i]*distance[i])){
                 currentPick=price[i];
-                ans+=(currentPick*distance[i]);
+                ans+=(currentPick *distance[i]);
             }else{
-                ans+=(currentPick*distance[i]);
+                ans+=(currentPick *distance[i]);
             }
         }
         System.out.println(ans);
